requirements.txt
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
mlflow>=2.9.0
pyyaml>=6.0.1
pytest>=7.4.0
flake8>=6.1.0